refactor: refactor handleUploadRequest into a standalone handler, instead of a uploadService function. refactor prehandlers into their own directory
