+++ runner/src/server/plugins/engine/plugin.ts
@@ -11,7 +11,11 @@ import { FormPayload } from "./types";
 import { shouldLogin } from "server/plugins/auth";
 import config from "../../config";
 import * as exit from "./pluginHandlers/exit";
-import { ReadableStreamEntry } from "server/services/upload/uploadService";
+import {
+  getFiles,
+  validateContentTypes,
+} from "./pluginHandlers/files/prehandlers";
+import { handleUploadRequest } from "server/plugins/engine/pluginHandlers/files/prehandlers/handleUploadRequest";
 
 configure([
   // Configure Nunjucks to allow rendering of content that is revealed conditionally.
@@ -281,15 +285,6 @@ export const plugin = {
 
     const { uploadService } = server.services([]);
 
-    const handleFiles = (request: HapiRequest, h: HapiResponseToolkit) => {
-      const { path, id } = request.params;
-      const model = forms[id];
-      const page = model?.pages.find(
-        (page) => normalisePath(page.path) === normalisePath(path)
-      );
-      return uploadService.handleUploadRequest(request, h, page.pageDef);
-    };
-
     //TODO:- Merge with POST /{id}/{path*} route, and move to ./pluginHandlers/id/*
     const postHandler = async (
       request: HapiRequest,
@@ -334,7 +329,7 @@ export const plugin = {
         pre: [
           { method: getFiles, assign: "files" },
           { method: validateContentTypes, assign: "validFiles" },
-          { method: handleFiles },
+          { method: handleUpload },
         ],
         handler: postHandler,
       },
@@ -347,79 +342,3 @@ export const plugin = {
     server.route(exit.statusGet);
   },
 };
-
-function getFiles(request: HapiRequest, _h: HapiResponseToolkit) {
-  const { uploadService } = request.services([]);
-  const files = uploadService.fileStreamsFromPayload(request.payload);
-  if (files.length) {
-    request.server.logger.info(
-      { id: request.yar.id, path: request.path },
-      `Found ${uploadService.fileSummary(files)} to process on ${request.path}`
-    );
-    return files;
-  }
-
-  return null;
-}
-
-async function validateContentTypes(
-  request: HapiRequest,
-  h: HapiResponseToolkit
-) {
-  const files = request.pre.files;
-
-  if (!files) {
-    return h.continue;
-  }
-
-  const { uploadService, cacheService } = request.services([]);
-  const logger = request.server.logger;
-  const loggerIdentifier = { id: request.yar.id, path: request.path };
-
-  const validFields: ReadableStreamEntry[] = [];
-  const erroredFields: string[] = [];
-  const { originalFilenames = {} } = await cacheService.getState(request);
-
-  for (const [fieldName, values] of files) {
-    const invalidFile = values.find(
-      (value) => !uploadService.validateContentType(value)
-    );
-    if (invalidFile) {
-      logger.error(
-        loggerIdentifier,
-        `User uploaded file with invalid content type or empty field for ${fieldName}, attempting to find previous upload`
-      );
-
-      const originalFilename = originalFilenames[fieldName];
-      if (!originalFilename) {
-        logger.error(
-          loggerIdentifier,
-          `User uploaded invalid content type or empty field for ${fieldName}, and has no previous upload for field. Deleting ${fieldName} from payload`
-        );
-        delete request.payload[fieldName];
-        erroredFields.push(fieldName);
-      }
-
-      if (originalFilename) {
-        logger.warn(
-          loggerIdentifier,
-          `User uploaded invalid content type or empty field for ${fieldName}, using existing upload ${originalFilename.location} instead`
-        );
-        request.payload[fieldName] = originalFilename.location;
-      }
-
-      continue;
-    }
-
-    validFields.push([fieldName, values]);
-  }
-
-  if (erroredFields) {
-    request.pre.hasInvalidContentTypes = true;
-    request.pre.errors = erroredFields.map((field) =>
-      uploadService.invalidFileTypeError(field)
-    );
-  }
-
-  return validFields;
-}
