+++ runner/src/server/services/upload/uploadService.ts
@@ -152,117 +152,6 @@ export class UploadService {
     return parsedError(fieldName, this.validFiletypesString);
   }
 
-  async handleUploadRequest(
-    request: HapiRequest,
-    h: HapiResponseToolkit,
-    page: any
-  ) {
-    const { cacheService } = request.services([]);
-    const state = await cacheService.getState(request);
-    let originalFilenames = state?.originalFilenames ?? {};
-    const files = request.pre.files;
-
-    if (!files) {
-      return h.continue;
-    }
-
-    /**
-     * files is an array of tuples containing key and value.
-     * value may be an array of file data where multiple files have been uploaded
-     */
-    const validFiles = request.pre.validFiles;
-
-    for (const entry of validFiles) {
-      const [fieldName, streams] = entry;
-      const loggerIdentifier = this.getLoggerIdentifier(request, { fieldName });
-      const previousUpload = originalFilenames[fieldName];
-
-      if (previousUpload) {
-        this.logger.info(
-          loggerIdentifier,
-          `User is attempting to overwrite ${fieldName} with location ${previousUpload.location}`
-        );
-      }
-
-      let response;
-
-      try {
-        response = await this.uploadDocuments(streams);
-      } catch (err) {
-        console.log("ERR", err);
-        if (err.data?.res) {
-          const { error } = this.parsedDocumentUploadResponse(err.data);
-          request.pre.errors = [
-            ...request.pre.errors,
-            parsedError(fieldName, error),
-          ];
-        } else if (err.code === "EPIPE") {
-          // ignore this error, it happens when the request is responded to by the doc upload service before the
-          // body has finished being sent. A valid response is still received.
-          this.logger.warn(
-            loggerIdentifier,
-            `Ignoring EPIPE response ${err.message}`
-          );
-        } else {
-          this.logger.error(
-            { ...loggerIdentifier, err },
-            `Error uploading document: ${err.message}`
-          );
-          request.pre.errors = [
-            ...(h.request.pre.errors || []),
-            parsedError(fieldName, err),
-          ];
-        }
-      }
-
-      const { location, warning, error } = response;
-
-      if (location) {
-        const originalFilename = streams
-          .map((stream) => stream.hapi?.filename)
-          .join(", ");
-
-        this.logger.info(
-          loggerIdentifier,
-          `Uploaded ${fieldName} successfully to ${location}`
-        );
-
-        originalFilenames[fieldName] = { location, originalFilename };
-        const {
-          originalFilenames: updatedFilenames,
-        } = await cacheService.mergeState(request, { originalFilenames });
-
-        this.logger.info(
-          { ...loggerIdentifier, allFiles: updatedFilenames },
-          `Updated originalFileNames for user`
-        );
-        request.payload[fieldName] = location;
-        originalFilenames = updatedFilenames;
-      }
-
-      if (warning) {
-        request.pre.warning = warning;
-        this.logger.warn(
-          loggerIdentifier,
-          `File was uploaded successfully but there was a warning ${warning}`
-        );
-      }
-
-      if (error) {
-        this.logger.error(
-          loggerIdentifier,
-          `Document upload API responded with an error ${error}`
-        );
-        request.pre.errors = [
-          ...(request.pre.errors || []),
-          parsedError(fieldName, error),
-        ];
-      }
-    }
-
-    return h.continue;
-  }
-
   downloadDocuments(paths: string[]) {
     const promises = paths.map((path) => get<string>(path, {}));
     return Promise.all(promises);
